feat: complete hybrid integration of collaborator frontend
- Integrated collaborator's dashboard and student views via 'collab' directory
- Migrated App.tsx to use CollabLayout inside ProtectedRoute
- Wired up AuthContext into collaborator's Sidebar and Topbar
- Resolved TypeScript/JSX TSConfig imports
- Implemented robust PyMuPDF + Gemini PDF question extractor script
- Ignored massive pdfdata directory in gitignore
- Streamlined backend route configurations for 6-tier educational models

+++ backend/app/models/attempt.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, DateTime, ForeignKey, Boolean
+from sqlalchemy import Column, Integer, String, DateTime, ForeignKey, Boolean, Float
 from sqlalchemy.orm import relationship
 from sqlalchemy.sql import func
 from app.database import Base
@@ -35,7 +35,7 @@ class MockTestAttempt(Base):
     started_at = Column(DateTime(timezone=True), server_default=func.now())
     completed_at = Column(DateTime(timezone=True), nullable=True)
     total_time_seconds = Column(Integer, nullable=True)  # Actual time taken
-    score = Column(Integer, default=0)
+    score = Column(Float, default=0.0)
     total_questions = Column(Integer, nullable=False)
     correct_answers = Column(Integer, default=0)
     incorrect_answers = Column(Integer, default=0)
